Basic form function - inputs, autocompletes, image uploads working - still need to normalize for DB
